fix: refactor Settings and add SubMenu
