+++ src/ui/globals/types.ts
@@ -22,7 +22,6 @@ enum OperationType {
   SCAN_CONNECTION = "scanConnection",
   ADD_CREDENTIAL = "addCredential",
   RECEIVE_CONNECTION = "receiveConnection",
-  SHOW_SETTINGS = "showSettings",
 }
 
 enum ToastMsgType {
