+++ src/ui/pages/Menu/Menu.types.ts
@@ -1,0 +1,11 @@
+import { ReactNode } from "react";
+
+interface SubMenuProps {
+  showSubMenu: boolean;
+  setShowSubMenu: (value: boolean) => void;
+  title: string;
+  additionalButtons?: ReactNode;
+  children: ReactNode;
+}
+
+export type { SubMenuProps };
