+++ src/ui/pages/Menu/components/Settings/Settings.scss
@@ -1,38 +1,8 @@
-.settings {
-  display: flex !important;
-  background: var(--ion-color-light-grey);
+.settings-content {
+  margin-top: 2rem;
 
-  &.show {
-    transform: translate3d(0, 0, 0);
-    transition-duration: 0.5s;
-  }
-
-  &.hide {
-    transform: translate3d(100%, 0, 0);
-    transition-duration: 0.5s;
-  }
-
-  ion-header ion-toolbar {
-    margin-top: 0.625rem;
-    --background: var(--ion-color-light-grey);
-
-    ion-title.md {
-      margin-left: -3rem;
-    }
-  }
-
-  ion-toolbar ion-buttons ion-button.close-button {
-    --background-activated: var(--ion-color-light-grey);
-    --background-focused: var(--ion-color-light-grey);
-    --background-hover: var(--ion-color-light-grey);
-  }
-
-  .settings-content {
-    margin-top: 2rem;
-
-    .settings-section-title {
-      margin: 0.625rem 1.25rem;
-    }
+  .settings-section-title {
+    margin: 0.625rem 1.25rem;
   }
 
   ion-card {
@@ -91,17 +61,7 @@
   }
 
   @media screen and (min-width: 250px) and (max-width: 370px) {
-    ion-header ion-toolbar {
-      margin-top: 0.531rem;
-
-      ion-title.md {
-        margin-left: -2rem;
-      }
-    }
-
-    .settings-content {
-      margin-top: 1rem;
-    }
+    margin-top: 1rem;
 
     ion-card {
       padding: 0.5rem 1rem;
