+++ src/ui/pages/Menu/components/Settings/Settings.test.tsx
@@ -1,14 +1,11 @@
 import { render } from "@testing-library/react";
 import { Provider } from "react-redux";
 import { Settings } from "./Settings";
-import { store } from "../../../store";
-import EN_TRANSLATIONS from "../../../locales/en/en.json";
-import { setCurrentOperation } from "../../../store/reducers/stateCache";
-import { OperationType } from "../../globals/types";
+import { store } from "../../../../../store";
+import EN_TRANSLATIONS from "../../../../../locales/en/en.json";
 
 describe("Settings page", () => {
   test("Renders Settings page", () => {
-    store.dispatch(setCurrentOperation(OperationType.SHOW_SETTINGS));
     const { getByTestId, getByText } = render(
       <Provider store={store}>
         <Settings />
